requirements.txt
fastapi
uvicorn
darts
langchain
langchain-groq
langchain-community
langchain-core
chromadb
python-dotenv
pandas
torch